% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alexseev}
\alias{alexseev}
\title{Example Data}
\format{data frame with 72 observations and 11 variables.
\describe{
 \item{region}{Region}
 \item{xenovote}{ldpr all vote 03}
 \item{slavicshare}{percent population slav, 1989}
 \item{slavicshare_changeonslav}{percentage-point change of the proportion non-slavic}
 \item{inc9903}{average income, change from 1999 to 2003}
 \item{eduhi02}{higher education, 2002}
 \item{unemp02}{unemployment, 2002}
 \item{apt9200}{privatized apartments from 1992 to 2003}
 \item{vsall03}{vote "against all" in 2003}
 \item{brdcont}{location along russia's borders with disputed areas}
}}
\source{
\url{http://mattgolder.com/files/research/jop2.zip}
}
\usage{
alexseev
}
\description{
Xenophobic Vote Share in 2003 Russian Duma Elections
}
\keyword{datasets}
