% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_marginal_effects.R
\name{extract_marginal_effects}
\alias{extract_marginal_effects}
\title{Extract Marginal Effect columns from a \dQuote{margins} object}
\usage{
extract_marginal_effects(x)
}
\arguments{
\item{x}{An object of class \dQuote{margins}, as returned by \code{\link{margins}}.}
}
\value{
A data.frame containing only those columns containing estimated marginal effects data.
}
\description{
Extract Marginal Effect columns from a \dQuote{margins} object
}
\examples{
require('datasets')
extract_marginal_effects(margins(lm(mpg ~ hp, data = mtcars[1:10,])))
}

