% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_internal.R
\name{marginal_effect}
\alias{marginal_effect}
\title{Marginal Effects Calculator}
\usage{
marginal_effect(x, data, atmeans = FALSE, type = c("response", "link",
  "terms"), vce = c("delta", "simulation", "bootstrap"), iterations = 50L,
  method = c("simple", "Richardson", "complex"), ...)
}
\arguments{
\item{x}{A model object.}

\item{data}{A data.frame over which to calculate marginal effects.}

\item{atmeans}{A logical indicating whether to estimate \dQuote{marginal effects at means} or \dQuote{average marginal effects} (the default).}

\item{type}{A character string indicating the type of marginal effects to estimate. Mostly relevant for non-linear models, where the reasonable options are \dQuote{response} (the default) or \dQuote{link} (i.e., on the scale of the linear predictor in a GLM).}

\item{vce}{A character string indicating the type of estimation procedure to use for estimating variances. The default (\dQuote{delta}) uses the delta method. Alternatives are \dQuote{bootstrap}, which uses bootstrap estimation, or \dQuote{simulation}, which averages across simulations drawn from the joint sampling distribution of model coefficients. The latter two are extremely time intensive.}

\item{iterations}{If \code{vce = "bootstrap"}, the number of bootstrap iterations. If \code{vce = "simulation"}, the number of simulated effects to draw. Ignored otherwise.}

\item{method}{A character string indicating the numeric derivative method to use when estimating marginal effects. See \code{\link[numDeriv]{grad}} for details.}

\item{\dots}{Ignored.}
}
\value{
A data.frame of class \dQuote{margins} containing the contents of \code{data}, fitted values for \code{x}, the standard errors of the fitted values, and any estimated marginal effects. This data.frame may have repeated column names (for the original variables and the margginal effects thereof). Marginal effects columns are distinguished by their class (\dQuote{marginaleffect}) and can be extracted using \code{\link{extract_marginal_effects}}. Attributes containing additional information, including the marginal effect variances and additional details.
}
\description{
This is the low-level marginal effects calculator called by \code{\link{margins}}.
}
\details{
Generally, it is not necessary to call this function directly because \code{\link{margins}} provides a simpler interface.

This is the internal, workhorse function that estimates marginal effects from model \code{x} for a data.frame, \code{data}, optionally estimating either \dQuote{average marginal effects} (when \code{atmeans = FALSE}) or \dQuote{marginal effects at means} (when \code{atmeans = TRUE}). In the former case, the marginal effects are estimated for each observation in the dataset and returned in full. In the latter, column means are taken for \code{data} and estimation is performed only these \dQuote{averaged} cases. The former is generally preferred because the latter may estimate marginal effects for cases that are unintuitive or not covered by the observed data (e.g., the effect when a binary variable in \code{data} is averaged to 0.6 rather than at 0 and 1, respectively).

To estimate marginal effects at specified values of \code{x} (other than means), use \code{\link{margins}} with the \code{at} parameter to specify values of covariates at which to estimate effects.

The choice of \code{vce} may be important. The default variance-covariance estimation procedure (\code{vce = "delta"}) uses the delta method to estimate marginal effect variances. This is the fastest method. When \code{vce = "simulation"}, coefficient estimates are repeatedly drawn from the asymptotic (multivariate normal) distribution of the model coefficients and each draw is used to estimate marginal effects, with the variance based upon the dispersion of those simulated effects. The number of interations used is given by \code{iterations}. For \code{vce = "bootstrap"}, the bootstrap is used to repeatedly subsample \code{data} and the variance of marginal effects is estimated from the variance of the bootstrap distribution. This method is markedly slower than the other two procedures and, obviously, it will probably fail if \code{atmeans = TRUE}. Again, \code{iterations} regulates the number of boostrap subsamples to draw.
}

