% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_effects.R
\name{marginal_effects}
\alias{marginal_effects}
\title{Differentiate a Model Object}
\usage{
marginal_effects(model, data, type = c("response", "link"),
  method = c("simple", "Richardson", "complex"))
}
\arguments{
\item{model}{A model object, perhaps returned by \code{\link[stats]{lm}} or \code{\link[stats]{glm}}}

\item{data}{A data.frame over which to calculate marginal effects.}

\item{type}{A character string indicating the type of marginal effects to estimate. Mostly relevant for non-linear models, where the reasonable options are \dQuote{response} (the default) or \dQuote{link} (i.e., on the scale of the linear predictor in a GLM).}

\item{method}{A character string indicating the numeric derivative method to use when estimating marginal effects. \dQuote{simple} optimizes for speed; \dQuote{Richardson} optimizes for accuracy. See \code{\link[numDeriv]{grad}} for details.}
}
\value{
An data.frame with dimensions equal to \code{data}, where each row is an observation and each column is the marginal effect of that variable for the data values provided by \code{data}.
}
\description{
Extract marginal effects via numerical differentiation from a model object, conditional on data
}
\details{
This function uses numeric differentiation (\code{\link[numDeriv]{grad}}) to extract marginal effects of an estimated model with respect to all variables specified in \code{data} and returns a data.frame containing the unit-specific marginal effects with respect to each variable included (or not included) in the model. (Note that this is not each \emph{coefficient}.)
}
\examples{
require("datasets")
x <- lm(mpg ~ cyl * hp + wt, data = mtcars)
marginal_effects(x)

}
\seealso{
\code{\link{margins}}, \code{\link{margins.lm}}, \code{\link{build_margins}}
}

