% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins-package.R
\docType{package}
\name{margins-package}
\alias{margins-package}
\title{Marginal effects for model objects}
\description{
This package is an R port of Stata's margins command, implemented as an S3 generic for model objects, like those of class \dQuote{lm} and \dQuote{glm}.
}
\author{
Thomas J. Leeper
}
\seealso{
\code{\link{margins}}, \code{\link{margins.lm}}, \code{\link{margins.glm}}
}
\keyword{package}

