% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_glm.R
\name{margins.glm}
\alias{margins.glm}
\title{Marginal Effects for Generalized Linear Models}
\usage{
\method{margins}{glm}(x, newdata, at = NULL, atmeans = FALSE, ...)
}
\arguments{
\item{x}{A model object of class \dQuote{glm}.}

\item{newdata}{A data.frame containing the data at which to evaluate the marginal effects, as in \code{\link[stats]{predict}}.}

\item{at}{A list of one or more named vectors, specifically values at which to calculate the marginal effects. See examples.}

\item{atmeans}{A logical indicating whether to calculate marginal effects at the means (i.e., partial effects at the average of all covariates), as opposed to the default average marginal effects (i.e., average partial effects), which is the default.}

\item{\dots}{Arguments passed to \code{\link{marginal_effect}}. One of particular relevance for GLMs is \code{type}.}
}
\value{
An object of class \dQuote{marginslist}, composed of one or more objects of class \dQuote{margins}.
}
\description{
Calculate marginal effects from estimated generalized linear models
}
\details{
Calculate marginal effects for a generalized linear model object.
}
\examples{
# basic examples
require("datasets")
x <- glm(mpg ~ cyl * hp + wt, data = mtcars)
margins(x, type = "response")
margins(x, type = "link")

}
\author{
Thomas J. Leeper
}
\references{
Greene, W.H. 2012. Econometric Analysis, 7th Ed. Boston: Pearson.

Stata manual: \code{margins}. Retrieved 2014-12-15 from \url{http://www.stata.com/manuals13/rmargins.pdf}.
}
\seealso{
\code{\link{margins.glm}}, \code{\link{plot.margins}}
}

