% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_lm.R
\name{margins.lm}
\alias{margins.lm}
\title{Marginal Effects for OLS Regression}
\usage{
\method{margins}{lm}(x, newdata = NULL, at = NULL, atmeans = FALSE,
  factors = "continuous", ...)
}
\arguments{
\item{x}{A model object of class \dQuote{lm}.}

\item{newdata}{A data.frame containing the data at which to evaluate the marginal effects, as in \code{predict}.}

\item{at}{A list of one or more named vectors, specifically values at which to calculate the marginal effects. See examples. NOTE: THIS DOES NOT CURRENTLY WORK WITH VARIABLES THAT ARE TRANSFORMED IN FORMULA EXPRESSIONS (e.g., \code{y~factor(x)} or \code{y~I(x^2)}).}

\item{atmeans}{A logical indicating whether to calculate marginal effects at the means (i.e., partial effects at the average), as opposed to the default average marginal effects (i.e., average partial effects), which is the default.}

\item{factors}{A character string indicating whether, if \dQuote{continuous} (the default), to calculate marginal effects treating factor variables as continuous or, if \dQuote{discrete}, to calculate marginal effects for factor variables as discrete first differences (i.e., the difference in fitted values when the factor is 1 versus 0).}

\item{\dots}{Arguments passed to \code{.margins}.}
}
\value{
An object of class \dQuote{margins}.
}
\description{
Calculate marginal effects for estimated models
}
\details{
Calculate marginal effects and associated standard errors for a model object.
}
\examples{
library("datasets")
x <- lm(mpg ~ cyl * hp + wt, data = mtcars)
(m <- margins(x)[[1]])
}
\author{
Thomas J. Leeper
}
\references{
Greene, W.H. 2012. Econometric Analysis, 7th Ed. Boston: Pearson.

Stata manual: \code{margins}. Retrieved 2014-12-15 from \url{http://www.stata.com/manuals13/rmargins.pdf}.
}

