% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_plm.R
\name{margins.plm}
\alias{margins.pglm}
\alias{margins.plm}
\title{Marginal Effects for Panel Regression Models}
\usage{
\method{margins}{plm}(model, data = NULL, at = NULL, ...)

\method{margins}{pglm}(model, data = NULL, at = NULL, ...)
}
\arguments{
\item{model}{A model object of class \dQuote{plm} or \dQuote{pglm}, from the \pkg{plm} package.}

\item{data}{A data.frame containing the data at which to evaluate the marginal effects, as in \code{\link[stats]{predict}}.}

\item{at}{A list of one or more named vectors, specifically values at which to calculate the marginal effects. See \code{\link{build_datalist}} for details on use.}

\item{\dots}{Arguments passed to \code{\link{marginal_effects}}. One of particular relevance for GLMs is \code{type}.}
}
\value{
An object of class \dQuote{marginslist}, composed of one or more objects of class \dQuote{margins}.
}
\description{
Calculate marginal effects from estimated panel linear and panel generalized linear models
}
\seealso{
\code{\link{plot.margins}}, \code{\link{extract_marginal_effects}}
}
\keyword{models}

