\name{margins_at}
\alias{margins_at}
\title{Calculate marginal effects for model objects at specified values}
\description{Calculate marginal effects for estimated models}
\usage{
margins_at(x, ...)
}
\arguments{
    \item{x}{}
    \item{atmeans}{}
    \item{at}{}
    \item{...}{}
}
\details{Calculate marginal effects and associated standard errors for a model object.}
\value{An object of class \dQuote{margins}.}
\author{Thomas J. Leeper}
%\references{}
%\seealso{}
%\examples{}