% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persp.R
\name{persp.lm}
\alias{persp.lm}
\title{Perspective plots for models}
\usage{
\method{persp}{lm}(x, which = attributes(terms(m))[["term.labels"]][1:2],
  z = c("prediction", "effect"), type = c("response", "link"), nx = 50L,
  ny = nx, theta = 45, phi = 10, shade = 0.75, xlab = which[1],
  ylab = which[2], zlab = if (match.arg(z) == "prediction") match.arg(z)
  else paste0(match.arg(z), " of ", which[1]), ...)
}
\arguments{
\item{x}{A model object.}

\item{which}{A two-element character vector specifying the names of variables to use as \samp{x} and \samp{y} predictor dimensions in the plot. See \code{\link[graphics]{persp}} for details.}

\item{z}{A character string specifying whether to draw \dQuote{prediction} (fitted values from the model, calculated using \code{\link[stats]{predict}}) or \dQuote{effect} (marginal effect of \code{which[1]}, using \code{\link{margins}}).}

\item{type}{A character string specifying whether to calculate predictions on the response scale (default) or link (only relevant for non-linear models).}

\item{nx}{An integer specifying the number of points across \code{x} at which to calculate the predicted value or marginal effect.}

\item{ny}{An integer specifying the number of points across \code{y} at which to calculate the predicted value or marginal effect.}

\item{theta}{An integer vector specifying the value of \code{theta} in \code{\link[graphics]{persp}}. If length greater than 1, multiple subplots are drawn with different rotations.}

\item{phi}{An integer vector specifying the value of \code{phi} in \code{\link[graphics]{persp}}. If length greater than 1, multiple subplots are drawn with different rotations.}

\item{shade}{An integer vector specifying the value of \code{shade} in \code{\link[graphics]{persp}}.}

\item{xlab}{A character string specifying the value of \code{xlab} in \code{\link[graphics]{persp}}.}

\item{ylab}{A character string specifying the value of \code{ylab} in \code{\link[graphics]{persp}}.}

\item{zlab}{A character string specifying the value of \code{zlab} in \code{\link[graphics]{persp}}.}

\item{\dots}{Additional arguments passed to \code{\link[graphics]{persp}}.}
}
\description{
Draw one or more perspectives plots reflecting predictions or marginal effects from a model.
}
\examples{
require('datasets')
m <- lm(mpg ~ wt*drat, data = mtcars)
persp(m, theta = c(45, 135, 225, 315))

m <- lm(mpg ~ wt*drat*I(drat^2), data = mtcars)
persp(m, c("drat", "wt"), z = "effect", nx = 10, ny = 10, ticktype = "detailed")
}

