% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.margins}
\alias{plot.margins}
\title{Plot Marginal Effects Estimates}
\usage{
\method{plot}{margins}(x, level = 0.95, term_map = NULL, ...)
}
\arguments{
\item{x}{An object of class \dQuote{margins}, as returned by
\code{\link{margins}}.}

\item{level}{A numeric value between 0 and 1 indicating the confidence level
to use when drawing error bars.}

\item{term_map}{named character vector. Values refer to the original
term/variable names. Names refer to the term names that will appear in
the plot. Terms which are omitted from this vector will be omitted from the
plot.}

\item{\dots}{Additional arguments passed to
\code{\link[ggplot2]{geom_pointrange}}, such as \code{shape}, \code{size},
\code{colour}, etc.}
}
\value{
The original \dQuote{margins} object \code{x}, invisibly.
}
\description{
An implementation of Stata's \samp{marginsplot} as an S3
  generic function
}
\details{
This function is invoked for its side effect: a basic dot plot with
error bars displaying marginal effects as generated by
\code{\link{margins}}, in the style of Stata's \samp{marginsplot} command.
}
\examples{
\dontrun{
  require("datasets")
  x <- lm(mpg ~ cyl * hp + wt, data = mtcars)
  mar <- margins(x)
  plot(mar)
}

}
\seealso{
\code{\link{margins}}, \code{\link{persp.lm}}
}
\keyword{graphics}
