\name{plot.margins}
\alias{plot.margins}
\title{Plot a margins object}
\description{Plot marginal effects}
\usage{
\method{plot}{margins}(x, at = 1:ncol(x$Effect), which = colnames(x$Effect), 
         labels = which, horiz = FALSE, xlab = "", ylab = "Marginal Effect",
         level = 0.95, pch = 21, points.col = "black", points.bg = "black",
         las = 0, cex = 1, lwd = 2, zeroline = TRUE, ...)
}
\arguments{
    \item{x}{An object of class \code{margins}.}
    \item{at}{A numeric vector specifying where to draw the marginal effects.}
    \item{which}{A character vector specifying the names of the factors whose marginal effects should be plotted.}
    \item{labels}{A character vector specifying the axis labels for the marginal effects.}
    \item{xlab}{A character string containing the x-axis (y-axis, if \code{horiz=TRUE}) label.}
    \item{ylab}{A character string containing the y-axis (x-axis, if \code{horiz=TRUE}) label.}
    \item{horiz}{A logical indicating whether the plot should be drawn horizontally.}
    \item{level}{A numeric vector specifying the confidence levels at which to draw the confidence bands for the marginal effects.}
    \item{pch}{See \code{?points}.}
    \item{points.col}{A character vector specifying the border color of the marginal effects point estimates. Vector recycling applies.}
    \item{points.bg}{A character vector specifying the fill color of the marginal effects point estimates. Vector recycling applies.}
    \item{las}{See \code{?par}.}
    \item{cex}{See \code{?par}.}
    \item{lwd}{A numeric vector specifying the width of confidence intervals. Vector recyling applies.}
    \item{zeroline}{A logical indicating whether to draw a line at x=0 on the plot.}
    \item{...}{Additional arguments.}
}
\details{This function is invoked for its side effect: a dot plot with error bars displaying marginal effects as generated by \code{\link{margins}}.}
\value{An object of class \dQuote{margins}.}
\author{Thomas J. Leeper}
%\references{}
%\seealso{}
\examples{
# plot average marginal effects
x <- glm(am ~ cyl + hp + wt, data = mtcars, family=binomial)
m <- margins(x)[[1]]

# marginal effects plot
plot(m, main = "Average Marginal Effects")

# horizontal orientation
plot(m, horiz = TRUE)

# multiple levels
plot(m, level = c(0.67, 0.95, 0.99), horiz = TRUE)
}