% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot}
\alias{cplot}
\alias{cplot.glm}
\alias{cplot.lm}
\title{Conditional predicted value and average marginal effect plots for models}
\usage{
cplot(object, ...)

\method{cplot}{lm}(object, x = attributes(terms(object))[["term.labels"]][1],
  dx = x, what = c("prediction", "effect"), type = c("response", "link"),
  data = object[["model"]], at, method = c("simple", "Richardson",
  "complex"), n = 25L, level = 0.95, draw = TRUE, xlab = x, ylab = if
  (match.arg(what) == "prediction") paste0("Predicted value") else
  paste0("Marginal effect of ", dx), xlim, ylim, lwd = 2, col = "black",
  lty = 1, se.type = c("lines", "shade"), se.col = "black",
  se.fill = grDevices::gray(0.5, 0.5), se.lwd = lwd, se.lty = if
  (match.arg(se.type) == "lines") 2 else 0, factor.pch = 19,
  factor.col = se.col, factor.fill = factor.col, factor.cex = 1,
  xaxs = "i", yaxs = xaxs, las = 1, rug = TRUE, rug.col = "black",
  rug.size = -0.02, ...)

\method{cplot}{glm}(object, x = attributes(terms(object))[["term.labels"]][1],
  dx = x, what = c("prediction", "effect"), type = c("response", "link"),
  data = object[["model"]], at, method = c("simple", "Richardson",
  "complex"), n = 25L, level = 0.95, draw = TRUE, xlab = x, ylab = if
  (match.arg(what) == "prediction") paste0("Predicted value") else
  paste0("Marginal effect of ", dx), xlim, ylim, lwd = 2, col = "black",
  lty = 1, se.type = c("lines", "shade"), se.col = "black",
  se.fill = grDevices::gray(0.5, 0.5), se.lwd = lwd, se.lty = if
  (match.arg(se.type) == "lines") 2 else 0, factor.pch = 19,
  factor.col = se.col, factor.fill = factor.col, factor.cex = 1,
  xaxs = "i", yaxs = xaxs, las = 1, rug = TRUE, rug.col = "black",
  rug.size = -0.02, ...)
}
\arguments{
\item{object}{A model object.}

\item{x}{A character string specifying the name of variable to use as the x-axis dimension in the plot.}

\item{dx}{If \code{what = "effect"}, the variable whose conditional marginal effect should be displayed. By default it is \code{x} (so the plot displays the marginal effect of \code{x} across values of \code{x}); ignored otherwise.}

\item{what}{A character string specifying whether to draw a \dQuote{prediction} (fitted values from the model, calculated using \code{\link[stats]{predict}}) or an \dQuote{effect} (average marginal effect of \code{dx} conditional on \code{x}, using \code{\link{margins}}).}

\item{type}{A character string specifying whether to calculate predictions on the response scale (default) or link (only relevant for non-linear models).}

\item{data}{A data frame to override the default value offered in \code{object[["model"]]}.}

\item{at}{Currently ignored.}

\item{method}{If \code{what = "effect"}, a character string indicating the numeric derivative method to use when estimating marginal effects. \dQuote{simple} optimizes for speed; \dQuote{Richardson} optimizes for accuracy. See \code{\link[numDeriv]{grad}} for details.}

\item{n}{An integer specifying the number of points across \code{x} at which to calculate the predicted value or marginal effect.}

\item{level}{The confidence level required (used to draw uncertainty bounds).}

\item{draw}{A logical (default \code{TRUE}), specifying whether to draw the plot. If \code{FALSE}, the data used in drawing are returned as a list of data.frames. This might be useful if you want to plot using an alternative plotting package (e.g., ggplot2).}

\item{xlab}{A character string specifying the value of \code{xlab} in \code{\link[graphics]{plot}}.}

\item{ylab}{A character string specifying the value of \code{ylab} in \code{\link[graphics]{plot}}.}

\item{xlim}{A two-element numeric vector specifying the x-axis limits. Set automatically if missing.}

\item{ylim}{A two-element numeric vector specifying the y-axis limits. Set automatically if missing.}

\item{lwd}{An integer specifying the width of the prediction or marginal effect line. See \code{\link[graphics]{lines}}. If \code{x} is a factor variable in the model, this is used to set the line width of the error bars.}

\item{col}{A character string specifying the color of the prediction or marginal effect line. If \code{x} is a factor variable in the model, this is used to set the color of the error bars.}

\item{lty}{An integer specifying the \dQuote{line type} of the prediction or marginal effect line. See \code{\link[graphics]{par}}. If \code{x} is a factor variable in the model, this is used to set the line type of the error bars.}

\item{se.type}{A character string specifying whether to draw the confidence interval as \dQuote{lines} (the default, using \code{\link[graphics]{lines}}) or a \dQuote{shade} (using \code{\link[graphics]{polygon}}).}

\item{se.col}{If \code{se.type = "lines"}, a character string specifying the color of the confidence interval lines. If \code{se.type = "shade"}, the color of the shaded region border.}

\item{se.fill}{If \code{se.type = "shade"}, the color of the shaded region. Ignored otherwise.}

\item{se.lwd}{If \code{se.type = "lines"}, the width of the confidence interval lines. See \code{\link[graphics]{lines}}.}

\item{se.lty}{If \code{se.type = "lines"}, an integer specifying the \dQuote{line type} of the confidence interval lines; if \code{se.type = "shade"}, the line type of the shaded polygon border. See \code{\link[graphics]{par}}.}

\item{factor.pch}{If \code{x} is a factor variable in the model, the shape to use when drawing points. See \code{\link[graphics]{points}}.}

\item{factor.col}{If \code{x} is a factor variable in the model, the color to use for the border of the points. See \code{\link[graphics]{points}}.}

\item{factor.fill}{If \code{x} is a factor variable in the model, the color to use for the fill of the points. See \code{\link[graphics]{points}}.}

\item{factor.cex}{If \code{x} is a factor variable in the model, the \dQuote{expansion factor} to use for the point size. See \code{\link[graphics]{points}}.}

\item{xaxs}{A character string specifying \code{xaxs}. See \code{\link[graphics]{par}}.}

\item{yaxs}{A character string specifying \code{xaxs}. See \code{\link[graphics]{par}}.}

\item{las}{An integer string specifying \code{las}. See \code{\link[graphics]{par}}.}

\item{rug}{A logical specifying whether to include an x-axis \dQuote{rug} (see \code{\link[graphics]{rug}}).}

\item{rug.col}{A character string specifying \code{col} to \code{\link[graphics]{rug}}.}

\item{rug.size}{A numeric value specifying \code{ticksize} to \code{\link[graphics]{rug}}.}

\item{\dots}{Additional arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
A tidy data.frame containing the data used to draw the plot.
}
\description{
Draw one or more conditioanl effects plots reflecting predictions or marginal effects from a model, conditional on a covariate. Currently methods exist for \dQuote{lm} and \dQuote{glm} models.
}
\examples{
\dontrun{
require('datasets')
# prediction from several angles
m <- lm(Sepal.Length ~ Sepal.Width, data = iris)
cplot(m)

# more complex model
m <- lm(Sepal.Length ~ Sepal.Width * Petal.Width * I(Petal.Width ^ 2), data = head(iris, 50))
## marginal effect of 'Petal.Width' across 'Petal.Width'
cplot(m, x = "Petal.Width", what = "effect", n = 10)

# marginal effect of 'Petal.Width' across 'Sepal.Width'
## without drawing the plot
## this might be useful for using, e.g., ggplot2 for plotting
tmp <- cplot(m, x = "Sepal.Width", dx = "Petal.Width", what = "effect", n = 10, draw = FALSE)
if (require("ggplot2")) {
    # use ggplot2 instead of base graphics
    ggplot(tmp, aes(x = Petal.Width, y = "effect")) + geom_line(lwd = 2) + 
      geom_line(aes(y = effect + 1.96*se.effect)) + geom_line(aes(y = effect - 1.96*se.effect))
}
# a non-linear model
m <- glm(am ~ wt*drat, data = mtcars, family = binomial)
cplot(m, x = "wt") # prediction

# effects on linear predictor and outcome
cplot(m, x = "drat", dx = "wt", what = "effect", type = "link")
cplot(m, x = "drat", dx = "wt", what = "effect", type = "response")

}
}
\seealso{
\code{\link{plot.margins}}, \code{\link{persp.lm}}
}
\keyword{graphics}
\keyword{hplot}

