% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot}
\alias{cplot}
\alias{cplot.glm}
\alias{cplot.lm}
\title{Conditional predicted value and marginal effect plots for models}
\usage{
cplot(object, ...)

\method{cplot}{lm}(object, x = attributes(terms(object))[["term.labels"]][1],
  dx = x, what = c("prediction", "effect"), type = c("response", "link"),
  n = 25L, level = 0.95, xlab = x, ylab = if (match.arg(what) ==
  "prediction") paste0("Predicted value") else paste0("Marginal effect of ",
  dx), xlim, ylim, lwd = 2, col = "black", lty = 1, se.type = c("lines",
  "shade"), se.col = "black", se.fill = "gray", se.lwd = 1, se.lty = 2,
  xaxs = "i", yaxs = xaxs, las = 1, rug = TRUE, rug.col = "black",
  rug.size = -0.02, ...)

\method{cplot}{glm}(object, x = attributes(terms(object))[["term.labels"]][1],
  dx = x, what = c("prediction", "effect"), type = c("response", "link"),
  n = 25L, level = 0.95, xlab = x, ylab = if (match.arg(what) ==
  "prediction") paste0("Predicted value") else paste0("Marginal effect of ",
  dx), xlim, ylim, lwd = 2, col = "black", lty = 1, se.type = c("lines",
  "shade"), se.col = "black", se.fill = "gray", se.lwd = 1, se.lty = 2,
  xaxs = "i", yaxs = xaxs, las = 1, rug = TRUE, rug.col = "black",
  rug.size = -0.02, ...)
}
\arguments{
\item{object}{A model object.}

\item{x}{A character string specifying the name of variable to use as the x-axis dimension in the plot.}

\item{dx}{If \code{what = "effect"}, the variable whose conditional marginal effect should be displayed. By default it is \code{x} (so the plot displays the marginal effect of \code{x} across values of \code{x}); ignored otherwise.}

\item{what}{A character string specifying whether to draw \dQuote{prediction} (fitted values from the model, calculated using \code{\link[stats]{predict}}) or \dQuote{effect} (marginal effect of \code{x}, using \code{\link{margins}}).}

\item{type}{A character string specifying whether to calculate predictions on the response scale (default) or link (only relevant for non-linear models).}

\item{n}{An integer specifying the number of points across \code{x} at which to calculate the predicted value or marginal effect.}

\item{level}{The confidence level required (used to draw uncertainty bounds).}

\item{xlab}{A character string specifying the value of \code{xlab} in \code{\link[graphics]{plot}}.}

\item{ylab}{A character string specifying the value of \code{ylab} in \code{\link[graphics]{plot}}.}

\item{xlim}{A two-element numeric vector specifying the x-axis limits. Set automatically if missing.}

\item{ylim}{A two-element numeric vector specifying the y-axis limits. Set automatically if missing.}

\item{lwd}{An integer specifying the width of the prediction or marginal effect line. See \code{\link[graphics]{lines}}.}

\item{col}{A character string specifying the color of the prediction or marginal effect line.}

\item{lty}{An integer specifying the \dQuote{line type} of the prediction or marginal effect line. See \code{\link[graphics]{par}}.}

\item{se.type}{A character string specifying whether to draw the confidence interval as \dQuote{lines} (the default, using \code{\link[graphics]{lines}}) or a \dQuote{shade} (using \code{\link[graphics]{polygon}}).}

\item{se.col}{If \code{se.type = "lines"}, a character string specifying the color of the confidence interval lines. If \code{se.type = "shade"}, the color of the shaded region border.}

\item{se.fill}{If \code{se.type = "shade"}, the color of the shaded region. Ignored otherwise.}

\item{se.lwd}{If \code{se.type = "lines"}, the width of the confidence interval lines. See \code{\link[graphics]{lines}}.}

\item{se.lty}{If \code{se.type = "lines"}, an integer specifying the \dQuote{line type} of the confidence interval lines; if \code{se.type = "shade"}, the line type of the shaded polygon border. See \code{\link[graphics]{par}}.}

\item{xaxs}{A character string specifying \code{xaxs}. See \code{\link[graphics]{par}}.}

\item{yaxs}{A character string specifying \code{xaxs}. See \code{\link[graphics]{par}}.}

\item{las}{An integer string specifying \code{las}. See \code{\link[graphics]{par}}.}

\item{rug}{A logical specifying whether to include an x-axis \dQuote{rug} (see \code{\link[graphics]{rug}}).}

\item{rug.col}{A character string specifying \code{col} to \code{\link[graphics]{rug}}.}

\item{rug.size}{A numeric value specifying \code{ticksize} to \code{\link[graphics]{rug}}.}

\item{\dots}{Additional arguments passed to \code{\link[graphics]{plot}}.}
}
\description{
Draw one or more conditioanl effects plots reflecting predictions or marginal effects from a model, conditional on a covariate. Currently methods exist for \dQuote{lm} and \dQuote{glm} models.
}
\details{
Currently, this implements \dQuote{marginal effects at means} of all covariates. And confidence intervals are not drawn for marginal effects plots.
}
\examples{
require('datasets')
# prediction from several angles
m <- lm(Sepal.Length ~ Sepal.Width, data = iris)
cplot(m)

# more complex model
m <- lm(Sepal.Length ~ Sepal.Width * Petal.Width * I(Petal.Width ^ 2), data = iris)
## marginal effect of 'Petal.Width' across 'Petal.Width'
cplot(m, x = "Petal.Width", what = "effect", n = 10)
## marginal effect of 'Petal.Width' across 'Sepal.Width'
cplot(m, x = "Sepal.Width", dx = "Petal.Width", what = "effect", n = 10)

# a non-linear model
m <- glm(am ~ wt*drat, data = mtcars, family = binomial)
cplot(m, x = "wt") # prediction

# effects on linear predictor and outcome
cplot(m, x = "drat", dx = "wt", what = "effect", type = "link")
cplot(m, x = "drat", dx = "wt", what = "effect", type = "response")

}
\seealso{
\code{\link{plot.margins}}, \code{\link{persp.lm}}
}

