% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplot.R
\name{cplot_extract.default}
\alias{cplot_extract.default}
\title{Internal function to extract data for `cplot`}
\usage{
\method{cplot_extract}{default}(object, data, dx, level, xvar, at, n, type,
  xvals, vcov, what, ...)
}
\arguments{
\item{level}{The confidence level required (used to draw uncertainty
bounds).}

\item{at}{Currently ignored.}

\item{n}{An integer specifying the number of points across \code{x} at which
to calculate the predicted value or marginal effect, when \code{x} is
numeric. Ignored otherwise.}

\item{xvals}{A numeric vector of values at which to calculate predictions or
marginal effects, if \code{x} is numeric. By default, it is calculated from
the data using \code{\link{seq_range}}. If \code{x} is a factor, this is
ignored, as is \code{n}.}

\item{...}{Additional arguments such as \code{colour}, \code{linetype},
\code{size}, \code{shape}, \code{fill}, \code{alpha}. These will be passed
to \code{ggplot2} geom functions to alter the style of the plot.  If `x` is
a factor, these arguments will be passed to
\code{\link[ggplot2]{geom_pointrange}}. If `x` is numeric, these arguments
will be passed to \code{\link[ggplot2]{geom_line}} and
\code{\link[ggplot2]{geom_ribbon}}.  The \code{alpha} and \code{fill}
arguments are not passed to \code{geom_line}. The \code{colour} argument is
not passed to \code{geom_ribbon}.}
}
\description{
Internal function to extract data for `cplot`
}
