% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_marginal_effects.R
\name{extract_marginal_effects}
\alias{extract_marginal_effects}
\alias{extract_marginal_effects.margins}
\alias{extract_marginal_effects.marginslist}
\title{Extract Marginal Effect columns from a \dQuote{margins} object}
\usage{
extract_marginal_effects(x, ...)

\method{extract_marginal_effects}{margins}(x, ...)

\method{extract_marginal_effects}{marginslist}(x, ...)
}
\arguments{
\item{x}{An object of class \dQuote{margins}, as returned by \code{\link{margins}}.}

\item{\dots}{Ignored.}
}
\value{
A data.frame containing only those columns containing estimated marginal effects data.
}
\description{
Extract Marginal Effect columns from a \dQuote{margins} object
}
\examples{
require('datasets')
m <- margins(lm(mpg ~ hp, data = mtcars[1:10,]))
extract_marginal_effects(m)
extract_marginal_effects(m[[1]])
}
\seealso{
\code{\link{build_margins}}, \code{\link{margins}}
}
\keyword{models}

