% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_internal.R
\name{marginal_effect}
\alias{marginal_effect}
\title{Marginal Effects Calculator}
\usage{
marginal_effect(x, data, atmeans = FALSE, type = c("response", "link",
  "terms"), vce = c("delta", "bootstrap", "simulation"), iterations = 500L,
  method = c("Richardson", "simple", "complex"), ...)
}
\arguments{
\item{x}{A model object.}

\item{data}{A data.frame. If missing, data are extracted from the \code{x}.}

\item{atmeans}{A logical indicating whether to estimate \dQuote{marginal effects at means} or \dQuote{average marginal effects} (the default).}

\item{type}{A character string indicating the type of marginal effects to estimate. Mostly relevant for non-linear models, where the reasonable options are \dQuote{response} (the default) or \dQuote{link} (i.e., on the scale of the linear predictor in a GLM).}

\item{vce}{A character string indicating the type of estimation procedure to use for estimating variances. The default (\dQuote{delta}) uses the delta method. Alternatives are \dQuote{bootstrap}, which uses bootstrap estimation, or \dQuote{simulation}, which averages across simulations drawn from the joint sampling distribution of model coefficients. The latter two are extremely time intensive.}

\item{iterations}{If \code{vce = "bootstrap"}, the number of bootstrap iterations. If \code{vce = "simulation"}, the number of simulated effects to draw. Ignored otherwise.}

\item{method}{A character string indicating the numeric derivative method to use when estimating marginal effects. See \code{\link[numDeriv]{grad}} for details.}

\item{\dots}{Ignored.}
}
\description{
This is the low-level marginal effects calculator called by \code{\link{margins}}.
}

