\name{margins-package}
\alias{margins-package}
\docType{package}
\title{Marginal effects for model objects}
\description{This package is an R port of Stata's margins command, implemented as an S3 generic for model objects, like those of class \dQuote{lm} and \dQuote{glm}.}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
%\references{}
\keyword{ package }
\seealso{\code{\link{margins}}}
%\examples{}
