% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins-package.R
\docType{package}
\name{margins-package}
\alias{margins-package}
\title{Marginal effects for model objects}
\description{
This package is an R port of Stata's \samp{margins} command, implemented as an S3 generic \code{margins()} for model objects, like those of class \dQuote{lm} and \dQuote{glm}. The function also provides a low-level function, \code{\link{marginal_effects}}, to estimate those quantities and return a data.frame of unit-specific effects; and a imports the low-level function \code{\link[prediction]{prediction}}, which provides a consistent data frame interface to \code{\link[stats]{predict}}.
}
\author{
Thomas J. Leeper
}
\seealso{
\code{\link{margins}}, \code{\link{marginal_effects}}, \code{\link[prediction]{prediction}}
}
\keyword{package}

