% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_lm.R
\name{margins.lm}
\alias{margins.lm}
\title{Marginal Effects for OLS Regression}
\usage{
\method{margins}{lm}(model, data, at = NULL, atmeans = FALSE, ...)
}
\arguments{
\item{model}{A model object of class \dQuote{lm}.}

\item{data}{A data.frame containing the data at which to evaluate the marginal effects, as in \code{\link[stats]{predict}}.}

\item{at}{A list of one or more named vectors, specifically values at which to calculate the marginal effects. See \code{\link{build_datalist}} for details on use.}

\item{atmeans}{A logical indicating whether to calculate marginal effects at the means (i.e., partial effects at the average of all covariates), as opposed to the default average marginal effects (i.e., average partial effects), which is the default.}

\item{\dots}{Arguments passed to \code{\link{marginal_effect}}.}
}
\value{
An object of class \dQuote{marginslist}, composed of one or more objects of class \dQuote{margins}.
}
\description{
Calculate marginal effects from estimated linear models
}
\details{
Calculates marginal effects for a linear model object using \code{\link{marginal_effects}} and returns a \dQuote{marginslist} list of one or more \dQuote{margins} objects.
}
\examples{
# basic examples
require("datasets")
x <- lm(mpg ~ cyl * hp + wt, data = mtcars)
margins(x)

}
\author{
Thomas J. Leeper
}
\references{
Greene, W.H. 2012. Econometric Analysis, 7th Ed. Boston: Pearson.

Stata manual: \code{margins}. Retrieved 2014-12-15 from \url{http://www.stata.com/manuals13/rmargins.pdf}.
}
\seealso{
\code{\link{margins.glm}}, \code{\link{plot.margins}}, \code{\link{extract_marginal_effects}}
}

