% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_plm.R
\name{margins.plm}
\alias{margins.pglm}
\alias{margins.plm}
\title{Marginal Effects for Panel Regression Models}
\usage{
\method{margins}{plm}(model, data = NULL, at = NULL, atmeans = FALSE, ...)

\method{margins}{pglm}(model, data = NULL, at = NULL, atmeans = FALSE,
  ...)
}
\arguments{
\item{model}{A model object of class \dQuote{plm} or \dQuote{pglm}, from the \pkg{plm} package.}

\item{data}{A data.frame containing the data at which to evaluate the marginal effects, as in \code{\link[stats]{predict}}.}

\item{at}{A list of one or more named vectors, specifically values at which to calculate the marginal effects. See \code{\link{build_datalist}} for details on use.}

\item{atmeans}{A logical indicating whether to calculate marginal effects at the means (i.e., partial effects at the average of all covariates), as opposed to the default average marginal effects (i.e., average partial effects), which is the default.}

\item{\dots}{Arguments passed to \code{\link{marginal_effects}}. One of particular relevance for GLMs is \code{type}.}
}
\description{
Calculate marginal effects from estimated panel linear and panel generalized linear models
}
\seealso{
\code{\link{margins.lm}}, \code{\link{margins.glm}}, \code{\link{plot.margins}}, \code{\link{extract_marginal_effects}}
}

