% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins}
\alias{margins}
\title{Marginal Effects Estimation}
\usage{
margins(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{\dots}{Addition arguments passed to methods.}
}
\description{
S3 generic function for estimating marginal effects from modelling objects
}
\seealso{
\code{\link{margins.lm}}, \code{\link{margins.glm}}
}

