\name{margins}
\alias{margins}
\alias{margins.lm}
\alias{margins.glm}
\alias{print.margins}
\title{Marginal effects for model objects}
\description{Calculate marginal effects for estimated models}
\usage{
margins(x, at = NULL, atmeans = FALSE, ...)
}
\arguments{
    \item{x}{}
    \item{atmeans}{}
    \item{at}{}
    \item{...}{}
}
\details{Calculate marginal effects and associated standard errors for a model object.}
\value{An object of class \dQuote{margins}.}
\author{Thomas J. Leeper}
%\references{}
%\seealso{}
%\examples{}