% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persp.R
\name{persp.lm}
\alias{persp.lm}
\title{Perspective plots for models}
\usage{
\method{persp}{lm}(x, xvar = attributes(terms(x))[["term.labels"]][1],
  yvar = attributes(terms(x))[["term.labels"]][2], what = c("prediction",
  "effect"), type = c("response", "link"), nx = 25L, ny = nx,
  theta = 45, phi = 10, shade = 0.75, xlab = xvar, ylab = yvar,
  zlab = if (match.arg(what) == "prediction") "Predicted value" else
  paste0("Marginal effect of ", xvar), ticktype = c("detailed", "simple"),
  ...)
}
\arguments{
\item{x}{A model object.}

\item{xvar}{A character string specifying the name of variable to use as the \samp{x} dimension in the plot. See \code{\link[graphics]{persp}} for details.}

\item{yvar}{A character string specifying the name of variable to use as the \samp{y} dimension in the plot. See \code{\link[graphics]{persp}} for details.}

\item{what}{A character string specifying whether to draw \dQuote{prediction} (fitted values from the model, calculated using \code{\link[stats]{predict}}) or \dQuote{effect} (marginal effect of \code{x}, using \code{\link{margins}}).}

\item{type}{A character string specifying whether to calculate predictions on the response scale (default) or link (only relevant for non-linear models).}

\item{nx}{An integer specifying the number of points across \code{x} at which to calculate the predicted value or marginal effect.}

\item{ny}{An integer specifying the number of points across \code{y} at which to calculate the predicted value or marginal effect.}

\item{theta}{An integer vector specifying the value of \code{theta} in \code{\link[graphics]{persp}}. If length greater than 1, multiple subplots are drawn with different rotations.}

\item{phi}{An integer vector specifying the value of \code{phi} in \code{\link[graphics]{persp}}. If length greater than 1, multiple subplots are drawn with different rotations.}

\item{shade}{An integer vector specifying the value of \code{shade} in \code{\link[graphics]{persp}}.}

\item{xlab}{A character string specifying the value of \code{xlab} in \code{\link[graphics]{persp}}.}

\item{ylab}{A character string specifying the value of \code{ylab} in \code{\link[graphics]{persp}}.}

\item{zlab}{A character string specifying the value of \code{zlab} (vertical axis label) in \code{\link[graphics]{persp}}.}

\item{ticktype}{A character string specifying one of: \dQuote{detailed} (the default) or \dQuote{simple}. See \code{\link[graphics]{persp}}.}

\item{\dots}{Additional arguments passed to \code{\link[graphics]{persp}}.}
}
\description{
Draw one or more perspectives plots reflecting predictions or marginal effects from a model. Currently methods exist for \dQuote{lm} and \dQuote{glm} models.
}
\details{
Currently, this implements \dQuote{marginal effects at means} of all covariates.
}
\examples{
require('datasets')
# prediction from several angles
m <- lm(mpg ~ wt*drat, data = mtcars)
persp(m, theta = c(45, 135, 225, 315))

# marginal effect of 'drat' across drat and wt
m <- lm(mpg ~ wt*drat*I(drat^2), data = mtcars)
persp(m, xvar = "drat", yvar = "wt", what = "effect", nx = 10, ny = 10, ticktype = "detailed")

# a non-linear model
m <- glm(am ~ wt*drat, data = mtcars, family = binomial)
persp(m, theta = c(30, 60)) # prediction

# effects on linear predictor and outcome
persp(m, xvar = "drat", yvar = "wt", what = "effect", type = "link")
persp(m, xvar = "drat", yvar = "wt", what = "effect", type = "response")

}
\seealso{
\code{\link{plot.margins}}, \code{\link{cplot}}
}

