\name{plot.margins}
\alias{plot.margins}
\title{Plot a margins object}
\description{Plot marginal effects}
\usage{
\method{plot}{margins}(x, at = 1:ncol(x$Effect), which = colnames(x$Effect), 
         labels = which, level = 0.95, pch = 21, points.col = "black", points.bg = "black",
         las = 0, cex = 1, lwd = 2, zeroline = TRUE, ...)
}
\arguments{
    \item{x}{}
    \item{at}{}
    \item{which}{}
    \item{labels}{}
    \item{level}{}
    \item{pch}{}
    \item{points.col}{}
    \item{points.bg}{}
    \itme{las}{}
    \item{cex}{}
    \item{lwd}{}
    \item{zeroline}{}
    \item{...}{}
}
\details{This function is invoked for its side effect: a dot plot with error bars displaying marginal effects as generated by \code{\link{margins}}.}
\value{An object of class \dQuote{margins}.}
\author{Thomas J. Leeper}
%\references{}
%\seealso{}
\examples{
# plot average marginal effects
x <- glm(am ~ cyl + hp + wt, data = mtcars, family=binomial)
m <- margins(x)

# marginal effects plot
plot(m, main = "Average Marginal Effects")

# horizontal orientation
plot(m, horiz = TRUE)

# multiple levels
plot(m, level = c(0.67, 0.95, 0.99), horiz = TRUE)
}