\name{plot.margins}
\alias{plot.margins}
\title{Plot a margins object}
\description{Plot marginal effects}
\usage{
\method{plot}{margins}(x, which = row.names(x), include.intercept = FALSE, 
         labels = which, bg = "black", lwd = 2, ...)
}
\arguments{
    \item{x}{}
    \item{which}{}
    \item(include.intercept}{}
    \item{labels}{}
    \item{bg}{}
    \item{lwd}{}
    \item{...}{}
}
\details{This function is invoked for its side effect: a dotchart with error bars displaying marginal effects as generated by \code{\link{margins}}.}
\value{An object of class \dQuote{margins}.}
\author{Thomas J. Leeper}
%\references{}
%\seealso{}
\examples{
x <- lm(grade ~ gpa + tuce + psi, data = d)
m <- margins(x)
plot(m)


x <- lm(grade ~ gpa + tuce + factor(psi), data = d)
m <- margins(x)
plot(m)
}