\name{plot.margins}
\alias{plot.margins}
\title{Plot a margins object}
\description{Plot marginal effects}
\usage{
\method{plot}{margins}(x, which = colnames(x$Effect), labels = which, 
         level = 0.95, bg = "black", lwd = 2, vline = NULL, ...)
}
\arguments{
    \item{x}{}
    \item{which}{}
    \item{labels}{}
    \item{level}{}
    \item{bg}{}
    \item{lwd}{}
    \item{vline}{}
    \item{...}{}
}
\details{This function is invoked for its side effect: a dotchart with error bars displaying marginal effects as generated by \code{\link{margins}}.}
\value{An object of class \dQuote{margins}.}
\author{Thomas J. Leeper}
%\references{}
%\seealso{}
\examples{
# plot average marginal effects
x <- glm(am ~ cyl + hp + wt, data = mtcars, family=binomial)
m <- margins(x)
plot(m)
}