% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction}
\alias{prediction}
\alias{prediction.glm}
\alias{prediction.lm}
\alias{prediction.loess}
\title{Extract Predictions from a Model Object}
\usage{
prediction(model, data, ...)

\method{prediction}{lm}(model, data, type = "response", ...)

\method{prediction}{glm}(model, data, type = c("response", "link"), ...)

\method{prediction}{loess}(model, data, type = "response", ...)
}
\arguments{
\item{model}{A model object, perhaps returned by \code{\link[stats]{lm}} or \code{\link[stats]{glm}}.}

\item{data}{A data.frame over which to calculate marginal effects.}

\item{type}{A character string indicating the type of marginal effects to estimate. Mostly relevant for non-linear models, where the reasonable options are \dQuote{response} (the default) or \dQuote{link} (i.e., on the scale of the linear predictor in a GLM).}

\item{\dots}{Additional arguments passed to \code{\link[stats]{predict}} methods.}
}
\value{
An data.frame with number of rows equal to number of rows in \code{data}, where each row is an observation and the two columns represent fitted/predicted values and the standard errors thereof.
require("datasets")
x <- lm(mpg ~ cyl * hp + wt, data = mtcars)
prediction(x)
}
\description{
Extract predicted values via \code{\link[stats]{predict}} from a model object, conditional on data
}
\details{
This function is simply a wrapper around \code{\link[stats]{predict}} that returns a data.frame containing predicted values with respect to all variables specified in \code{data}. It is used internally by \code{\link{build_margins}}.
}
\seealso{
\code{\link{marginal_effects}}, \code{\link{build_margins}}
}
\keyword{models}

