% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{prediction}
\alias{find_data}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{prediction}{\code{\link[prediction]{find_data}}, \code{\link[prediction]{prediction}}}
}}

